Able to delete document

+++ src/app/pages/dashboard/document-management/services/document-management.service.ts
@@ -1,11 +1,14 @@
 import { Injectable } from "@angular/core";
 import { QueryFn } from "@angular/fire/compat/firestore";
 import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
+import * as moment from "moment";
 import { BehaviorSubject, Observable } from "rxjs";
 import { filterByEqual } from "src/app/common/utils/common-util";
 import { getMoment } from "src/app/common/utils/moment.util";
+import { takeOnce } from "src/app/common/utils/rxjs-util";
 import { DataService } from "src/app/core/services/data-service";
 import { FireStorageService } from "src/app/core/services/fire-storage.service";
+import { ToastService } from "src/app/core/toast/toast.service";
 import { DocumentUploadModalComponent } from "../components/document-upload-modal/document-upload-modal.component";
 import { Document, DocumentDto, DocumentSearch } from "../models/document.model";
 
@@ -19,6 +22,7 @@ export class DocumentManagementService {
 
     constructor(private dataService: DataService,
         private storageService: FireStorageService,
+        private toastService: ToastService,
         private modalService: NgbModal) {
     }
 
@@ -30,13 +34,14 @@ export class DocumentManagementService {
         const criteriaQuery: QueryFn = ref => {
             let query: firebase.default.firestore.CollectionReference | firebase.default.firestore.Query = ref;
             if (criteria.uploadBy) query = query.where('meta.createdBy', '==', criteria.uploadBy);
-            if (criteria.startDate) query = query.where('meta.createdDate', '>=', criteria.startDate);
-            if (criteria.endDate) query = query.where('meta.createdDate', '<=', criteria.endDate);
-            return ref.orderBy('meta.createdDate', 'desc');
+            if (criteria.startDate) query = query.where('meta.createdDate', '>=', moment(criteria.startDate).startOf('day').toDate());
+            if (criteria.endDate) query = query.where('meta.createdDate', '<=', moment(criteria.endDate).endOf('day').toDate());
+            return query.orderBy('meta.createdDate', 'desc');
         };
 
         let collection = await this.dataService.getCollection<Document>(DocumentManagementService.DOCUMENT_COLLECTION_PATH, { showSpinner: true, query: criteriaQuery });
         if (criteria.name) collection = collection.filter(filterByEqual('name', criteria.name));
+        if (!collection[0]) this.toastService.showSuccess('No data found');
         this.document$.next(collection);
     }
 
@@ -53,9 +58,14 @@ export class DocumentManagementService {
     }
 
     async deleteDocument(documentDto: DocumentDto): Promise<void> {
+        const confirmation = confirm('After confirm the content will be deleted from the system.');
+        if (!confirmation) return;
+
         await this.storageService.deleteFile(documentDto.path);
         await this.dataService.deleteDocument(DocumentManagementService.DOCUMENT_COLLECTION_PATH, documentDto.meta.documentId!,
             { showSpinner: true, toastMessage: 'Document deleted' });
+        this.document$.pipe(takeOnce()).subscribe(documents =>
+            this.document$.next(documents.filter(doc => doc.meta.documentId !== documentDto.meta.documentId)));
     }
 
     clearSearchResult(): void {
